requirements.txt
greenlet==1.1.3
numpy==1.23.2
pandas==1.4.4
PyMySQL==1.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.40
